# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "fast_track — crypto price tracker",
    "version": "0.0.0",
    "description": "Tracks cryptocurrency prices from CoinMarketCap, stores the history in SQLite and syncs prices in the background.\n\nAll `/api` endpoints need an API key: `Authorization: Bearer ft_...`. Keys are created with `npm run apikey:create`. Roles: `read` (read endpoints) and `admin` (everything)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "listJobs",
        "summary": "Background jobs and their last run (admin)",
        "responses": {
          "200": {
            "description": "Job summary",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummaryList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/jobs/runs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "listJobRuns",
        "summary": "Job run history, newest first (admin)",
        "parameters": [
          {
            "name": "job",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sync-prices"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/JobRunStatus"
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Page of runs",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRunList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ft_<key>",
        "description": "Service API key, for example `ft_...`. Not the CoinMarketCap key."
      }
    },
    "headers": {
      "XRequestId": {
        "description": "Request id, echoed from the request or generated",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "parameters": {
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
...
    },
    "schemas": {
...
      "JobRunStatus": {
        "type": "string",
        "enum": [
          "running",
          "success",
          "partial",
          "failed",
          "skipped",
          "aborted"
        ]
      },
      "JobRun": {
        "type": "object",
        "required": [
          "id",
          "jobName",
          "status",
          "startedAt",
          "finishedAt",
          "durationMs",
          "itemsTotal",
          "itemsOk",
          "itemsFailed",
          "errorCode",
          "errorMessage"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "jobName": {
            "type": "string",
            "example": "sync-prices"
          },
          "status": {
            "$ref": "#/components/schemas/JobRunStatus"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "durationMs": {
            "type": "integer",
            "nullable": true
          },
          "itemsTotal": {
            "type": "integer"
          },
          "itemsOk": {
            "type": "integer"
          },
          "itemsFailed": {
            "type": "integer"
          },
          "errorCode": {
            "type": "string",
            "nullable": true,
            "example": "CMC_RATE_LIMITED"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "JobRunList": {
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobRun"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
      "JobSummary": {
        "type": "object",
        "required": [
          "name",
          "enabled",
          "running",
          "nextRunAt",
          "consecutiveFailures",
          "intervalMs",
          "lastRun"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "running": {
            "type": "boolean"
          },
          "nextRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "consecutiveFailures": {
            "type": "integer"
          },
          "intervalMs": {
            "type": "integer"
          },
          "lastRun": {
            "allOf": [
              {
                "$ref": "#/components/schemas/JobRun"
              }
            ],
            "nullable": true
          }
        }
      },
      "JobSummaryList": {
        "type": "object",
        "required": [
          "jobs"
        ],
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            }
          }
        }
      },
      "ValidationDetail": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationDetail"
                }
              }
            }
          }
        },
        "example": {
          "error": {
            "code": "VALIDATION_ERROR",
            "message": "Invalid request",
            "requestId": "3f1c...",
            "details": [
              {
                "field": "limit",
                "message": "must be an integer 1..100"
              }
            ]
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input (VALIDATION_ERROR)",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, unknown or revoked API key",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          },
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "example": "Bearer"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The key role does not allow this operation",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}